Create README.md
Pizza Sales PowerBI Dashboard We used here Power BI for Data Visulization. To Fill KPI Requirement of Pizza Shop: Which Pizza, Category, and Size Affects Sales We Build Bar and Funnel charts to show daily and monthly trends on Total Order and Total Revenue. And able to find out Best Selling and Worst Selling Pizza. Conclude that the Highest orders are spotted on Friday and Saturday in July Month and Brie Carre Pizza is the Worst Seller and Classic Pizza is the Best Seller.